refactor: checklist component subfolders (#4038)
