+++ editor.planx.uk/src/@planx/components/Checklist/Public/Public.tsx
@@ -2,18 +2,16 @@ import Box from "@mui/material/Box";
 import Grid from "@mui/material/Grid";
 import { visuallyHidden } from "@mui/utils";
 import {
-  type Checklist,
   checklistValidationSchema,
   getFlatOptions,
   getLayout,
-  type Group,
 } from "@planx/components/Checklist/model";
 import ImageButton from "@planx/components/shared/Buttons/ImageButton";
 import Card from "@planx/components/shared/Preview/Card";
 import { CardHeader } from "@planx/components/shared/Preview/CardHeader/CardHeader";
 import { getIn, useFormik } from "formik";
 import { useStore } from "pages/FlowEditor/lib/store";
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { ExpandableList, ExpandableListItem } from "ui/public/ExpandableList";
 import FormWrapper from "ui/public/FormWrapper";
 import FullWidthWrapper from "ui/public/FullWidthWrapper";
@@ -21,21 +19,14 @@ import ChecklistItem from "ui/shared/ChecklistItem/ChecklistItem";
 import ErrorWrapper from "ui/shared/ErrorWrapper";
 import { object } from "yup";
 
-import { Option } from "../shared";
-import type { PublicProps } from "../shared/types";
-
-export type Props = PublicProps<Checklist>;
+import { Props } from "../types";
+import { AutoAnsweredChecklist } from "./AutoAnsweredChecklist";
+import { getInitialExpandedGroups, toggleInArray } from "./helpers";
 
 export enum ChecklistLayout {
   Basic,
   Grouped,
   Images,
-}
-
-function toggleInArray<T>(value: T, arr: Array<T>): Array<T> {
-  return arr.includes(value)
-    ? arr.filter((val) => val !== value)
-    : [...arr, value];
 }
 
 const ChecklistComponent: React.FC<Props> = (props) => {
@@ -56,20 +47,6 @@ const ChecklistComponent: React.FC<Props> = (props) => {
   }
 
   return <VisibleChecklist {...props} />;
-};
-
-// An auto-answered Checklist won't be seen by the user, but still leaves a breadcrumb
-const AutoAnsweredChecklist: React.FC<Props & { answerIds: string[] }> = (
-  props,
-) => {
-  useEffect(() => {
-    props.handleSubmit?.({
-      answers: props.answerIds,
-      auto: true,
-    });
-  }, []);
-
-  return null;
 };
 
 const VisibleChecklist: React.FC<Props> = (props) => {
@@ -101,7 +78,10 @@ const VisibleChecklist: React.FC<Props> = (props) => {
     }),
   });
 
-  const initialExpandedGroups = getInitialExpandedGroups();
+  const initialExpandedGroups = getInitialExpandedGroups(
+    groupedOptions,
+    previouslySubmittedData,
+  );
 
   const [expandedGroups, setExpandedGroups] = useState<Array<number>>(
     initialExpandedGroups,
@@ -111,13 +91,9 @@ const VisibleChecklist: React.FC<Props> = (props) => {
   const flatOptions = getFlatOptions({ options, groupedOptions });
 
   const changeCheckbox = (id: string) => (_checked: any) => {
-    let newCheckedIds;
-
-    if (formik.values.checked.includes(id)) {
-      newCheckedIds = formik.values.checked.filter((x) => x !== id);
-    } else {
-      newCheckedIds = [...formik.values.checked, id];
-    }
+    const newCheckedIds = formik.values.checked.includes(id)
+      ? formik.values.checked.filter((x) => x !== id)
+      : [...formik.values.checked, id];
 
     formik.setFieldValue(
       "checked",
@@ -146,38 +122,37 @@ const VisibleChecklist: React.FC<Props> = (props) => {
             component="fieldset"
           >
             <legend style={visuallyHidden}>{text}</legend>
-            {options &&
-              options.map((option) =>
-                layout === ChecklistLayout.Basic ? (
-                  <FormWrapper key={option.id}>
-                    <Grid item xs={12} key={option.data.text}>
-                      <ChecklistItem
-                        onChange={changeCheckbox(option.id)}
-                        label={option.data.text}
-                        id={option.id}
-                        checked={formik.values.checked.includes(option.id)}
-                      />
-                    </Grid>
-                  </FormWrapper>
-                ) : (
-                  <Grid
-                    item
-                    xs={12}
-                    sm={6}
-                    contentWrap={4}
-                    key={option.data.text}
-                  >
-                    <ImageButton
-                      title={option.data.text}
+            {options?.map((option) =>
+              layout === ChecklistLayout.Basic ? (
+                <FormWrapper key={option.id}>
+                  <Grid item xs={12} key={option.data.text}>
+                    <ChecklistItem
+                      onChange={changeCheckbox(option.id)}
+                      label={option.data.text}
                       id={option.id}
-                      img={option.data.img}
-                      selected={formik.values.checked.includes(option.id)}
-                      onClick={changeCheckbox(option.id)}
-                      checkbox
+                      checked={formik.values.checked.includes(option.id)}
                     />
                   </Grid>
-                ),
-              )}
+                </FormWrapper>
+              ) : (
+                <Grid
+                  item
+                  xs={12}
+                  sm={6}
+                  contentWrap={4}
+                  key={option.data.text}
+                >
+                  <ImageButton
+                    title={option.data.text}
+                    id={option.id}
+                    img={option.data.img}
+                    selected={formik.values.checked.includes(option.id)}
+                    onClick={changeCheckbox(option.id)}
+                    checkbox
+                  />
+                </Grid>
+              ),
+            )}
 
             {groupedOptions && (
               <FormWrapper>
@@ -231,21 +206,5 @@ const VisibleChecklist: React.FC<Props> = (props) => {
       </FullWidthWrapper>
     </Card>
   );
-
-  function getInitialExpandedGroups() {
-    return (groupedOptions ?? ([] as Group<Option>[])).reduce(
-      (acc, group, index) =>
-        groupHasOptionSelected(group, previouslySubmittedData?.answers ?? [])
-          ? [...acc, index]
-          : acc,
-      [] as number[],
-    );
-
-    function groupHasOptionSelected(group: Group<Option>, answers: string[]) {
-      return group.children.some((child) =>
-        answers.some((id) => child.id === id),
-      );
-    }
-  }
 };
 export default ChecklistComponent;
